Caching all room messages...
Currently all room messages are stored in rooms.js with the time that
cache will increase too much, so we need to put a limit there and don't
store more than a certain number of messages.

That's dificult because the room messages pagination in matrix, so we
need to find a way to load more messages from a message id, not from a
batch token.