Get the room avatar after sync
Some rooms doesn't have avatar setted and we need to calculate wich
avatar to use. For example direct chat rooms should show the other
member avatar, and that calc is done in the backend, so we need to call
the GetRoomAvatar command to get the correct avatar.

This will make the avatar to change in the initial sync, but is the
easier way that I found to have the correct avatar displayed in the room
list.