Speed up message loading: Load always last 40 msgs
We're loading all messages stored in memory every time, and with rooms
with a lot of activity that will cause a huge time to render all
messages when changing from one room to another.

To solve that, I'm loading only 40 messages, no matter how many there
are in the memory. If the user clicks on load more, messages are getted
from cache and if there's no more, we call to the API.