Global AppOp to avoid gtk loops
With a global AppOp reference we can launch backend loop and internal
loop in a rust thread so now we can lock until a message comes from the
channel. With this change we don't need to poll the channel.

The APPOP macro runs the AppOp method in the glib main thread, using
this global variable.