Going to chat when reloading
The join room needs some work, currently is working, but we're reloading
all rooms with a full sync, that takes time.

It will be better to add the new room to the room list manually, because
we've the information from the directory and then get the room members
and room messages.