Load more messages recursive to avoid zero messages
We're ignoring membership events when loading more messages, so if
there's a lot of leave/enter events, we get 0 messages when we try to
load more. This patch retry always gets more messages, because if
there's only no text events, a recursive call is made to find at least
one.