Ability to launch multiple Fractal instances
Using the FRACTAL_ID environment variable the app will be lauch with
that id so it's possible to have multiple instances running.

The problem with this is that we're using the same folder for the cache
and secret-service so it's not possible yet to use this functionality to
open different accounts.

See #96