password storage functions move to passwd mod
This is a first step to split app.rs in different files.

I've implemented as a trait because this was the easier way, with this
change we don't need to change anything in the app.rs because this trait
adds the needed methods to the struct.

I want to add other password storage methods, for the people that
doesn't have secret-service, maybe configurable as a gsettings.