SecretService storage as a different module
This patch moves the password and token store to a different module and
let the trait methods clean. Using this as a base it's possible to
implement other password storage modules and add some logic in the trait
to choose which method to use. So we can use an environment variable or
gsettings to choose an insecure password storage, like a file in cache
directory or maybe use other password storage implementations.