Vscrollbar always to load more without button
If we remove the button to load more, we need to show always the
vertical scrollbar to receive the edge-reached event.

There's a accesibility problem here, because I've not been able to
launch this event without the mouse wheel so currently there's no way to
load more with the keyboard or only the mouse without the wheel, as far
as I now.

Fix #48