Fix autocomplete crash with unicode usernames
Sometimes, when we try to autocomplete with usernames with unicodes, the
count of chars is different from the count of graphemes so we were
trying to split the list of graphemes at a out of bound position.

Making the check first and returning if it's not possible fixs the
problem.

Fix #202