autocomplete: Code cleanup, remove all unwraps
I've removed all unwraps that can be removed, managing the result of all
of that with the ?  operator and using the std::option::Option type.

This way we'll avoid crashes, because an unmanaged unwrap can cause a
crash so we need to check this always.

I've also removed the unicode-segmentation depencency. We can use the
.chars iterators and it seems to work correctly.