Message: Implement PartialOrd trait
This gives a manual implementation of the PartialOrd trait for Message
which is a chronological order. We first check that the messages are
not the same in order to be sure that no time offset could make the
comparison erroneous, then we simply compare the dates and times of
the messages.