Fix EMPTY ROOM for 1:1 bug
This patch fix the problem that causes the elimination of member events
from the initial sync.

To calculate 1:1 rooms we use the members in that rooms so if we don't
have that information we should ask to the matrix server.

I've just move some code to the fractal-gtk to calculate the room name
when we receive the list of member in a room. Meantime we'll show a
"..." that I suppose that any user will understand as "loading".

This loading is done in the background so the initial sync works faster
and if we've rooms without name we'll ask to the server about the
members and then we'll change the room name according to the number of
members.