stickers: Stickers popover and send
This patch adds a new button to the main window to show the stickers. A
popover is shown when you click in this new button with all the stickers
group.

If you've purchased the stickers you'll view all the stickers there and
you can click over one sticker to send it.

Stickers not purchased are shown as a preview thumbnail and description
and a button to add is there too but not working yet.