stickers: Purchase sticker group working
The purchase is working but the sticker popover refresh doesn't work
well, the stickers doesn't update just after the purchase.

There's also other problem when there's a lot of stickers in a group we
try to load all at once and that's a problem, we should load one by one
or something like that to avoid huge resource consumption.