directory: Spinner for loading
I've added an spinner meantime we're loading rooms in the directory.
I've also changed the backend to return each third party once it's
finished so we don't need to wait for all third party, we'll view
results as soon as possible.

I've to move the ordering of rooms to the frontend.

I've also added a default image in the image widget.