directory: Disable search entry when loading
I've disabled the search entry when we're search for rooms to avoid
multiple requests. Right now you can't cancel a search request.

I've also moved the initial request to the set_protocols method so we
don't call the search rooms every time we enter in the room directory,
the room directory will be the same if you go back to chat and go to
directory again.

see #250