api: Improve threading performance of API
There are two backend commands that were blocking the rest of requests,
the GetUserInfoAsync and the DirectorySearch. To fix this I've moved the
slow code to a new thread so the backend can continue receiving
commands.

I've also added a timeout param to the get! and post! macros so we can
define a custom timeout for directory search queries because that takes
a lot of time.