Add a message queue and retry send on failure
I've added a message queue for new messages to keep the list of temp
messages, this messages are sent one by one and that means that we'll
preserve the order always.

If a send message fails the app tries to resend the same message again
and don't try to send any new message util the first one is sent
correctly.

Fix #201, #173 and #140